refactor: move event_manager and event_instances under interfaces/events/
