+++ axis/interfaces/events/__init__.py
@@ -1,0 +1,1 @@
+"""Event interface handlers for Axis devices."""
