+++ axis/interfaces/events/event_manager.py
@@ -4,8 +4,8 @@ from collections.abc import Callable
 import logging
 from typing import Any
 
-from ..models.event import Event, EventOperation, EventTopic
-from ..models.events.topic_normalizer import to_canonical
+from ...models.event import Event, EventOperation, EventTopic
+from ...models.events.topic_normalizer import to_canonical
 
 SubscriptionCallback = Callable[[Event], None]
 SubscriptionType = tuple[
